Colors constant created Routes moved from types to constants
