+++ src/constants/Colors.ts
@@ -1,0 +1,6 @@
+import Colors from '~/styles/export/colors.scss';
+
+export const Color = {
+    accent: Colors.accent,
+    active: Colors.active,
+};
