+++ src/constants/index.ts
@@ -1,0 +1,2 @@
+export { Routes } from './Routes';
+export { Color } from './Colors';
