+++ src/types/index.ts
@@ -1,2 +1,1 @@
 export { default as SVGIconProps } from './Icon';
-export { Routes } from './Routes';
